+++ lib/providers/services/navigation_service.dart
@@ -19,12 +19,12 @@ import 'package:seeds/screens/app/profile/image_viewer.dart';
 import 'package:seeds/screens/app/profile/logout.dart';
 import 'package:seeds/screens/app/wallet/custom_transaction.dart';
 import 'package:seeds/screens/app/wallet/dashboard.dart';
-import 'package:seeds/screens/app/wallet/receive.dart';
-import 'package:seeds/screens/app/wallet/receive_qr.dart';
+import 'package:seeds/screens/app/wallet/receive/receive.dart';
+import 'package:seeds/screens/app/wallet/receive/receive_qr.dart';
 import 'package:seeds/screens/app/wallet/scan.dart';
-import 'package:seeds/screens/app/wallet/transfer.dart';
-import 'package:seeds/screens/app/wallet/transfer_amount.dart';
-import 'package:seeds/screens/app/wallet/transfer_form.dart';
+import 'package:seeds/screens/app/wallet/transfer/transfer.dart';
+import 'package:seeds/screens/app/wallet/transfer/transfer_amount.dart';
+import 'package:seeds/screens/app/wallet/transfer/transfer_form.dart';
 import 'package:seeds/screens/onboarding/join_process.dart';
 import 'package:seeds/screens/onboarding/onboarding.dart';
 import 'package:seeds/widgets/page_not_found.dart';
