+++ lib/screens/app/wallet/receive/receive_qr.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart' hide Action;
 import 'package:seeds/providers/services/eos_service.dart';
 import 'package:seeds/widgets/main_button.dart';
-import '../../../utils/double_extension.dart';
+import '../../../../utils/double_extension.dart';
 import 'package:qr_flutter/qr_flutter.dart';
 import 'package:seeds/i18n/wallet.i18n.dart';
